add nft mint page
